+++ src/jquery.fancytree.js
@@ -1969,7 +1969,8 @@ FancytreeNode.prototype = /** @lends FancytreeNode# */{
 		return this.tree._callHook("nodeToggleSelected", this);
 	},
 	toString: function() {
-		return "<FancytreeNode(#" + this.key + ", '" + this.title + "')>";
+		return "FancytreeNode@" + this.key + "[title='" + this.title + "']";
+		// return "<FancytreeNode(#" + this.key + ", '" + this.title + "')>";
 	},
 	/**
 	 * Trigger `modifyChild` event on a parent to signal that a child was modified.
@@ -2864,7 +2865,8 @@ Fancytree.prototype = /** @lends Fancytree# */{
 	 * @returns {string}
 	 */
 	toString: function(){
-		return "<Fancytree(#" + this._id + ")>";
+		return "Fancytree@" + this._id;
+		// return "<Fancytree(#" + this._id + ")>";
 	},
 	/* _trigger a widget event with additional node ctx.
 	 * @see EventData